chore: move casting logic into module
